Conflicts:
	include/fieldmap.h
	include/menu.h
	include/text.h
	src/battle_controller_player.c
	src/battle_gfx_sfx_util.c
	src/battle_pyramid.c
	src/fieldmap.c
	src/item_use.c
	src/menu.c
	src/text.c

+++ src/secret_base.c
@@ -1283,7 +1283,7 @@ void SecretBasePerStepCallback(u8 taskId)
             PopSecretBaseBalloon(MapGridGetMetatileIdAt(x, y), x, y);
             if (sInFriendSecretBase == TRUE)
             {
-                switch ((int)MapGridGetMetatileIdAt(x, y))
+                switch (MapGridGetMetatileIdAt(x, y))
                 {
                 case METATILE_SecretBase_RedBalloon:
                 case METATILE_SecretBase_BlueBalloon:
@@ -2065,7 +2065,7 @@ void CheckInteractedWithFriendsSandOrnament(void)
     s16 x, y;
 
     GetXYCoordsOneStepInFrontOfPlayer(&x, &y);
-    switch ((int)MapGridGetMetatileIdAt(x, y))
+    switch (MapGridGetMetatileIdAt(x, y))
     {
     case METATILE_SecretBase_SandOrnament_Base1:
     case METATILE_SecretBase_SandOrnament_Base2:
